+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Enhancement: Read tracesSampleRate from AndroidManifest
 * Fix: Initialize Logback after context refreshes (#1129)
 * Ref: Return NoOpTransaction instead of null (#1126)
+* Fix: Do not crash when passing null values to @Nullable methods, eg User and Scope
 
 # 4.0.0-alpha.2